Refactor reminders tab into modular components
