+++ src/components/goals/RemindersTab.tsx
@@ -1,163 +1,19 @@
 // src/components/goals/RemindersTab.tsx
 "use client";
 
-/**
- * RemindersTab — Hero header + borderless TabBar everywhere
- * - Domain sub-tabs in Hero (Life | League | Learn), neon divider tint matches domain
- * - Bottom search uses built-in Hero search
- * - Quick Add row now lives inside the SAME panel as the cards (top of SectionCard.Body)
- * - Groups row uses TabBar (badges show per-group counts)
- * - Filters panel (toggle): Source (TabBar) + Pinned chip
- *
- * Notes:
- * - Removed ad-hoc CSS import ("../goals/style.css") to keep globals as source of truth.
- * - Button/IconButton use canonical props; delete uses tone="danger".
- * - Typings added to onChange handlers to avoid implicit any.
- */
-
 import * as React from "react";
 import SectionCard from "@/components/ui/layout/SectionCard";
-import Input from "@/components/ui/primitives/Input";
-import Textarea from "@/components/ui/primitives/Textarea";
 import Button from "@/components/ui/primitives/Button";
-import IconButton from "@/components/ui/primitives/IconButton";
 import Hero from "@/components/ui/layout/Hero";
-import TabBar from "@/components/ui/layout/TabBar";
-import SegmentedButton from "@/components/ui/primitives/SegmentedButton";
-import { uid, usePersistentState } from "@/lib/db";
-import useAutoFocus from "@/lib/useAutoFocus";
+import { Search, Sparkles, Gamepad2, GraduationCap, Plus } from "lucide-react";
+import ReminderQuickAddForm from "./reminders/ReminderQuickAddForm";
+import ReminderFilters from "./reminders/ReminderFilters";
+import ReminderList from "./reminders/ReminderList";
 import {
-  Search,
-  SlidersHorizontal,
-  Pin,
-  PinOff,
-  Trash2,
-  Sparkles,
-  Gamepad2,
-  GraduationCap,
-  Pencil,
-  Plus,
-} from "lucide-react";
-
-/* ───────── Types & seeds ───────── */
-
-type Domain = "Life" | "League" | "Learn";
-type Source = "MLA" | "BLA" | "BrokenByConcept" | "Custom";
-type Group = "quick" | "pregame" | "laning" | "trading" | "tempo" | "review";
-
-type Reminder = {
-  id: string;
-  title: string;
-  body?: string;
-  tags: string[];
-  source: Source;
-  group: Group;
-  pinned?: boolean;
-  domain?: Domain;
-  createdAt: number;
-  updatedAt: number;
-};
-
-const STORE_KEY = "goals.reminders.v4";
-
-function R(
-  title: string,
-  body: string,
-  tags: string[],
-  source: Source,
-  group: Group,
-  pinned = false,
-  domain?: Domain,
-): Reminder {
-  const now = Date.now();
-  return {
-    id: uid("rem"),
-    title,
-    body,
-    tags,
-    source,
-    group,
-    pinned,
-    domain,
-    createdAt: now,
-    updatedAt: now,
-  };
-}
-
-const SEEDS: Reminder[] = [
-  R(
-    "Hit 2 first",
-    "Push 3 melees → step up or respect if losing push.",
-    [],
-    "BLA",
-    "quick",
-    true,
-    "Learn",
-  ),
-  R(
-    "Jungle start check",
-    "Track start → first gank lane → second spawn prio.",
-    [],
-    "BrokenByConcept",
-    "quick",
-    true,
-    "Learn",
-  ),
-  R(
-    "3-wave plan",
-    "Say it: slow then crash 3 OR hold then thin/freeze.",
-    [],
-    "MLA",
-    "pregame",
-    false,
-    "League",
-  ),
-  R(
-    "Ward 2:30",
-    "River/tri at 2:30. Sweep before shove.",
-    [],
-    "BrokenByConcept",
-    "pregame",
-    false,
-    "Learn",
-  ),
-  R(
-    "Space with casters",
-    "Trade when your casters live; back off on enemy wave.",
-    [],
-    "MLA",
-    "laning",
-    false,
-    "League",
-  ),
-  R(
-    "CD punish",
-    "Trade on enemy cooldown gaps; track sums.",
-    [],
-    "MLA",
-    "trading",
-    false,
-    "League",
-  ),
-  R(
-    "Good recall",
-    "Shove → recall on spike → arrive first to river.",
-    [],
-    "BrokenByConcept",
-    "tempo",
-    false,
-    "Learn",
-  ),
-  R(
-    "Death audit",
-    "Wave, vision, jungle, greed. Name the fix.",
-    [],
-    "BLA",
-    "review",
-    false,
-    "Learn",
-  ),
-];
+  RemindersProvider,
+  useReminders,
+  Domain,
+} from "./reminders/useReminders";
 
 const DOMAIN_ITEMS: Array<{
   key: Domain;
@@ -169,128 +25,16 @@ const DOMAIN_ITEMS: Array<{
   { key: "Learn", label: "Learn", icon: <GraduationCap className="mr-1" /> },
 ];
 
-const GROUPS: Array<{ key: Group; label: string; hint?: string }> = [
-  { key: "quick", label: "Quick", hint: "Pin-worthy" },
-  { key: "pregame", label: "Pre-Game", hint: "Before queue" },
-  { key: "laning", label: "Laning", hint: "Wave & trades" },
-  { key: "trading", label: "Trading", hint: "Windows & sums" },
-  { key: "tempo", label: "Tempo", hint: "Recall & prio" },
-  { key: "review", label: "Review", hint: "After game" },
-];
-
-/* ───────── Utils ───────── */
-
-const pad = (n: number) => (n < 10 ? `0${n}` : `${n}`);
-const fmtDate = (ts: number) => {
-  const d = new Date(ts);
-  return `${pad(d.getDate())}/${pad(d.getMonth() + 1)}/${d.getFullYear()}`;
-};
-const cap = (s: string) => s.slice(0, 1).toUpperCase() + s.slice(1);
-
-/* ───────── Page ───────── */
-
 export default function RemindersTab() {
-  const [items, setItems] = usePersistentState<Reminder[]>(STORE_KEY, SEEDS);
-  const [query, setQuery] = React.useState("");
-  const [onlyPinned, setOnlyPinned] = React.useState(false);
-  const [domain, setDomain] = usePersistentState<Domain>(
-    "goals.reminders.domain.v2",
-    "League",
+  return (
+    <RemindersProvider>
+      <RemindersContent />
+    </RemindersProvider>
   );
-  const [group, setGroup] = usePersistentState<Group>(
-    "goals.reminders.group.v1",
-    "quick",
-  );
-  const [source, setSource] = usePersistentState<Source | "all">(
-    "goals.reminders.source.v1",
-    "all",
-  );
-  const [quickAdd, setQuickAdd] = React.useState("");
-  const [showFilters, setShowFilters] = React.useState(false);
+}
 
-  const inferDomain = (r: Reminder): Domain =>
-    r.domain ??
-    (r.source === "BLA" || r.source === "BrokenByConcept" ? "Learn" : "League");
-
-  // counts for group badges (per current domain)
-  const counts = React.useMemo(() => {
-    const domItems = items.filter((r) => inferDomain(r) === domain);
-    return domItems.reduce(
-      (acc, r) => ((acc[r.group] = (acc[r.group] ?? 0) + 1), acc),
-      {
-        quick: 0,
-        pregame: 0,
-        laning: 0,
-        trading: 0,
-        tempo: 0,
-        review: 0,
-      } as Record<Group, number>,
-    );
-  }, [items, domain]);
-
-  const filtered = React.useMemo(() => {
-    const q = query.trim().toLowerCase();
-    return items
-      .filter((r) => inferDomain(r) === domain)
-      .filter((r) => r.group === group)
-      .filter((r) => (source === "all" ? true : r.source === source))
-      .filter((r) => (onlyPinned ? r.pinned : true))
-      .filter((r) => {
-        if (!q) return true;
-        const hay = [r.title, r.body ?? "", r.tags.join(" "), r.source, r.group]
-          .join(" ")
-          .toLowerCase();
-        return hay.includes(q);
-      })
-      .sort((a, b) => {
-        if (!!b.pinned === !!a.pinned) return b.updatedAt - a.updatedAt;
-        return Number(b.pinned) - Number(a.pinned);
-      });
-  }, [items, domain, group, source, onlyPinned, query]);
-
-  function addNew(initialTitle?: string) {
-    const now = Date.now();
-    const next: Reminder = {
-      id: uid("rem"),
-      title: (initialTitle ?? "New reminder").trim() || "New reminder",
-      body: "",
-      tags: [],
-      source: domain === "Learn" ? "BLA" : "Custom",
-      group,
-      domain,
-      pinned: group === "quick",
-      createdAt: now,
-      updatedAt: now,
-    };
-    setItems((prev) => [next, ...prev]);
-    setQuickAdd("");
-  }
-
-  const update = (id: string, partial: Partial<Reminder>) =>
-    setItems((prev) =>
-      prev.map((r) =>
-        r.id === id ? { ...r, ...partial, updatedAt: Date.now() } : r,
-      ),
-    );
-
-  const remove = (id: string) =>
-    setItems((prev) => prev.filter((r) => r.id !== id));
-
-  const showGroups = domain === "League" || domain === "Learn";
-  const neonClass = domain === "Life" ? "neon-life" : "neon-primary";
-
-  // TabBar items
-  const GROUP_TABS = GROUPS.map((g) => ({
-    key: g.key,
-    label: g.label,
-    badge: counts[g.key], // numeric pill
-  }));
-  const SOURCE_TABS = (
-    ["all", "MLA", "BLA", "BrokenByConcept", "Custom"] as Array<"all" | Source>
-  ).map((s) => ({
-    key: s,
-    label: s === "all" ? "All" : s,
-  }));
+function RemindersContent() {
+  const { domain, setDomain, query, setQuery, filtered, addReminder } = useReminders();
 
   return (
     <>
@@ -304,7 +48,7 @@ export default function RemindersTab() {
         subTabs={{
           items: DOMAIN_ITEMS,
           value: domain,
-          onChange: (k: Domain) => setDomain(k),
+          onChange: (key: Domain) => setDomain(key),
           align: "end",
           size: "md",
           ariaLabel: "Reminder domain",
@@ -317,7 +61,9 @@ export default function RemindersTab() {
           debounceMs: 80,
           right: (
             <div className="flex items-center gap-2">
-              <span className="text-label font-medium tracking-[0.02em] opacity-75">{filtered.length}</span>
+              <span className="text-label font-medium tracking-[0.02em] opacity-75">
+                {filtered.length}
+              </span>
               <Search className="opacity-80" size={16} />
             </div>
           ),
@@ -327,7 +73,7 @@ export default function RemindersTab() {
             variant="primary"
             size="md"
             className="px-4 whitespace-nowrap"
-            onClick={() => addNew()}
+            onClick={() => addReminder()}
           >
             <Plus />
             <span>New Reminder</span>
@@ -338,401 +84,13 @@ export default function RemindersTab() {
       <SectionCard className="goal-card">
         <SectionCard.Body>
           <div className="grid gap-3">
-            {/* Quick Add row — now INSIDE the same panel as the cards */}
-            <form
-              onSubmit={(e) => {
-                e.preventDefault();
-                if (quickAdd.trim()) addNew(quickAdd);
-              }}
-              className="rounded-card flex items-center gap-2 sm:gap-6 glitch"
-            >
-              <Input
-                aria-label="Quick add reminder"
-                placeholder={`Quick add to ${GROUPS.find((g) => g.key === group)?.label ?? "Group"}…`}
-                value={quickAdd}
-                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                  setQuickAdd(e.currentTarget.value)
-                }
-                className="flex-1"
-              />
-              <IconButton
-                title="Add quick"
-                aria-label="Add quick"
-                type="submit"
-                size="md"
-                variant="solid"
-              >
-                <Plus size={16} aria-hidden />
-              </IconButton>
-              <div className={`${neonClass} hidden sm:block`}>
-                <p className="neon-note neon-glow text-label font-medium tracking-[0.02em] italic">
-                  Stop procrastinating, do it now if you have time
-                </p>
-              </div>
-            </form>
-
-            {/* Groups row (League/Learn) with Filters toggle on the right */}
-            {showGroups && (
-              <TabBar
-                items={GROUP_TABS}
-                value={group}
-                onValueChange={(k) => setGroup(k as Group)}
-                ariaLabel="Reminder group"
-                size="md"
-                align="between"
-                className="overflow-x-auto"
-                right={
-                  <SegmentedButton
-                    className="inline-flex items-center gap-1"
-                    onClick={() => setShowFilters((v) => !v)}
-                    aria-expanded={showFilters}
-                    title="Filters"
-                    isActive={showFilters}
-                  >
-                    <SlidersHorizontal size={16} aria-hidden />
-                    Filters
-                  </SegmentedButton>
-                }
-              />
-            )}
-
-            {/* Filters panel (collapsible) */}
-            {showFilters && (
-              <div className="flex flex-wrap items-center gap-4 pl-1">
-                <TabBar
-                  items={SOURCE_TABS}
-                  value={source}
-                  onValueChange={(k) => setSource(k as Source | "all")}
-                  ariaLabel="Reminder source filter"
-                  size="sm"
-                />
-                <SegmentedButton
-                  onClick={() => setOnlyPinned((v) => !v)}
-                  aria-pressed={onlyPinned}
-                  title="Pinned only"
-                  isActive={onlyPinned}
-                >
-                  {onlyPinned ? (
-                    <PinOff className="mr-1" />
-                  ) : (
-                    <Pin className="mr-1" />
-                  )}
-                  {onlyPinned ? "Pinned only" : "Any pin"}
-                </SegmentedButton>
-              </div>
-            )}
-
-            {/* Cards grid */}
-            <div className="grid grid-cols-12 gap-3">
-              {filtered.map((r) => (
-                <div
-                  key={r.id}
-                  className="col-span-12 md:col-span-6 lg:col-span-4"
-                >
-                  <RemTile
-                    value={r}
-                    onChange={(p) => update(r.id, p)}
-                    onDelete={() => remove(r.id)}
-                  />
-                </div>
-              ))}
-            </div>
-
-            {filtered.length === 0 && <EmptyState />}
+            <ReminderQuickAddForm />
+            <ReminderFilters />
+            <ReminderList />
           </div>
         </SectionCard.Body>
       </SectionCard>
-
-      {/* Local styles: keep neon-note flicker; divider is handled by Hero */}
-      <style jsx>{`
-        .neon-primary {
-          --neon: var(--primary);
-        }
-        .neon-life {
-          --neon: var(--accent);
-        }
-
-        .neon-note {
-          margin-top: calc(var(--space-1) * -1.5);
-          padding-left: calc(var(--space-1) / 2);
-          animation: neon-flicker 4s infinite;
-        }
-
-        @keyframes neon-flicker {
-          0%,
-          17%,
-          22%,
-          26%,
-          52%,
-          100% {
-            opacity: 1;
-          }
-          18% {
-            opacity: 0.72;
-          }
-          24% {
-            opacity: 0.55;
-          }
-          54% {
-            opacity: 0.78;
-          }
-          70% {
-            opacity: 0.62;
-          }
-          74% {
-            opacity: 1;
-          }
-        }
-
-        @media (prefers-reduced-motion: reduce) {
-          .neon-note {
-            animation: none;
-          }
-        }
-      `}</style>
     </>
   );
 }
 
-/* ───────── UI bits ───────── */
-
-function EmptyState() {
-  return (
-    <div className="goal-card rounded-card ds-card-pad text-ui font-medium text-muted-foreground grid place-items-center">
-      <p>Nothing here. Add one clear sentence you’ll read in champ select.</p>
-    </div>
-  );
-}
-
-/* ───────── Tile (goals-style, globals-driven) ───────── */
-
-function RemTile({
-  value,
-  onChange,
-  onDelete,
-}: {
-  value: Reminder;
-  onChange: (p: Partial<Reminder>) => void;
-  onDelete: () => void;
-}) {
-  const [editing, setEditing] = React.useState(false);
-  const [title, setTitle] = React.useState(value.title);
-  const [body, setBody] = React.useState(value.body ?? "");
-  const [tagsText, setTagsText] = React.useState(value.tags.join(", "));
-  const titleRef = React.useRef<HTMLInputElement | null>(null);
-
-  useAutoFocus({ ref: titleRef, when: editing });
-
-  function save() {
-    const cleanTags = tagsText
-      .split(",")
-      .map((t) => t.trim())
-      .filter(Boolean);
-    onChange({
-      title: title.trim() || "Untitled",
-      body: body.trim(),
-      tags: cleanTags,
-    });
-    setEditing(false);
-  }
-
-  const pinned = !!value.pinned;
-
-  return (
-    <article className="card-neo rounded-card card-pad relative group">
-      {/* Header row */}
-      <div className="flex items-center justify-between gap-2">
-        <div className="flex-1 min-w-0">
-          {editing ? (
-            <Input
-              ref={titleRef}
-              value={title}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                setTitle(e.currentTarget.value)
-              }
-              onKeyDown={(e: React.KeyboardEvent<HTMLInputElement>) =>
-                e.key === "Enter" && save()
-              }
-              aria-label="Title"
-              className="font-semibold uppercase tracking-wide"
-            />
-          ) : (
-            <div className="flex items-center gap-2 min-w-0">
-              <h4
-                className="font-semibold uppercase tracking-wide pr-2 title-glow glitch leading-6 truncate"
-                title={value.title}
-              >
-                {value.title}
-              </h4>
-              <Button
-                size="sm"
-                variant="ghost"
-                onClick={() => {
-                  setEditing(true);
-                }}
-                aria-label={value.title ? `Edit reminder ${value.title}` : "Edit reminder"}
-                className="shrink-0"
-              >
-                <Pencil aria-hidden="true" />
-                Edit
-              </Button>
-            </div>
-          )}
-        </div>
-
-        <div className="flex items-center gap-1">
-          <IconButton
-            title="Delete"
-            aria-label="Delete"
-            onClick={onDelete}
-            size="sm"
-            iconSize="sm"
-            variant="ring"
-            className="opacity-0 pointer-events-none group-hover:opacity-100 group-hover:pointer-events-auto transition-opacity"
-          >
-            <Trash2 />
-          </IconButton>
-
-          <IconButton
-            aria-pressed={pinned}
-            title={pinned ? "Unpin" : "Pin"}
-            aria-label={pinned ? "Unpin" : "Pin"}
-            onClick={() => onChange({ pinned: !pinned })}
-            size="sm"
-            iconSize="sm"
-          >
-            {pinned ? <PinOff /> : <Pin />}
-          </IconButton>
-        </div>
-      </div>
-
-      {/* Body + meta */}
-      <div className="mt-2 space-y-3">
-        {editing ? (
-          <>
-            <label className="text-label font-medium tracking-[0.02em] opacity-70">Note</label>
-            <Textarea
-              aria-label="Body"
-              placeholder="Short, skimmable sentence."
-              value={body}
-              onChange={(e: React.ChangeEvent<HTMLTextAreaElement>) =>
-                setBody(e.currentTarget.value)
-              }
-              className="rounded-[var(--radius-2xl)]"
-              resize="resize-y"
-              textareaClassName="min-h-40 leading-relaxed"
-            />
-
-            <label className="text-label font-medium tracking-[0.02em] opacity-70">Tags</label>
-            <Input
-              aria-label="Tags (comma separated)"
-              placeholder="tags, comma, separated"
-              value={tagsText}
-              onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                setTagsText(e.currentTarget.value)
-              }
-            />
-
-            <div className="segmented flex flex-wrap -m-1">
-              {(
-                [
-                  "pregame",
-                  "laning",
-                  "trading",
-                  "tempo",
-                  "review",
-                  "quick",
-                ] as Group[]
-              ).map((g) => (
-                <SegmentedButton
-                  key={g}
-                  onClick={() => onChange({ group: g })}
-                  isActive={value.group === g}
-                  className="m-1"
-                >
-                  {g === "pregame" ? "Pre-Game" : cap(g)}
-                </SegmentedButton>
-              ))}
-            </div>
-
-            <div className="segmented flex flex-wrap -m-1">
-              {(["MLA", "BLA", "BrokenByConcept", "Custom"] as Source[]).map(
-                (s) => (
-                  <SegmentedButton
-                    key={s}
-                    onClick={() => onChange({ source: s })}
-                    isActive={value.source === s}
-                    className="m-1"
-                  >
-                    {s}
-                  </SegmentedButton>
-                ),
-              )}
-            </div>
-
-            <div className="segmented flex flex-wrap -m-1">
-              {(["Life", "League", "Learn"] as Domain[]).map((d) => (
-                <SegmentedButton
-                  key={d}
-                  onClick={() => onChange({ domain: d })}
-                  isActive={(value.domain ?? "League") === d}
-                  className="m-1"
-                >
-                  {d}
-                </SegmentedButton>
-              ))}
-            </div>
-
-            <div className="flex gap-2">
-              <Button size="sm" onClick={save}>
-                Save
-              </Button>
-              <Button
-                size="sm"
-                variant="ghost"
-                onClick={() => {
-                  setEditing(false);
-                  setTitle(value.title);
-                  setBody(value.body ?? "");
-                  setTagsText(value.tags.join(", "));
-                }}
-              >
-                Cancel
-              </Button>
-            </div>
-          </>
-        ) : (
-          <>
-            <p className="text-ui font-medium">
-              <span className="opacity-70">Note:</span>{" "}
-              {value.body || (
-                <span className="opacity-60">
-                  No text. Click title to edit.
-                </span>
-              )}
-            </p>
-
-            <div className="mt-1 flex items-center justify-between text-ui font-medium">
-              <div className="flex items-center gap-2">
-                <span
-                  aria-hidden="true"
-                  className="inline-block h-2 w-2 rounded-full bg-[hsl(var(--accent-overlay))]"
-                />
-                <span className="text-label font-medium tracking-[0.02em]">{fmtDate(value.createdAt)}</span>
-              </div>
-
-              <button
-                className="text-label font-medium tracking-[0.02em] underline underline-offset-2 text-primary hover:brightness-75"
-                onClick={() => onChange({ pinned: !pinned })}
-                title={pinned ? "Unpin" : "Pin"}
-                type="button"
-              >
-                {pinned ? "Pinned" : "Pin"}
-              </button>
-            </div>
-          </>
-        )}
-      </div>
-    </article>
-  );
-}
