+++ src/components/ui/layout/PageHeader.tsx
@@ -82,7 +82,7 @@ const PageHeaderInner = <
   }: PageHeaderBaseProps<HeaderKey, HeroKey>,
   ref: React.ForwardedRef<PageHeaderFrameElement>,
 ) => {
-  const Component = (as ?? "header") as PageHeaderElement;
+  const Component = (as ?? "section") as PageHeaderElement;
 
   const {
     subTabs: heroSubTabs,
